+++ DesignerMacros/src/services-cqrs-crud/create-crud-macro-advanced-mapping/create-crud-macro-advanced-mapping.ts
@@ -8,6 +8,14 @@ async function execute(element: IElementApi, domainClass?: IElementApi) {
         return;
     }
 
+    if (privateSettersOnly && !hasConstructor(entity)) {
+        await dialogService.error(
+`Unable to create CQRS Operations.
+
+Private Setters are enabled with no constructor present on entity '${entity.getName()}'. In order to create CQRS Operations for that entity, either disable private setters or model a constructor element and try again.`);
+        return;
+    }
+
     const owningEntity = DomainHelper.getOwningAggregate(entity);
     const folderName = pluralize(DomainHelper.ownerIsAggregateRoot(entity) ? owningEntity.getName() : entity.getName());
     const folder = element.getChildren().find(x => x.getName() == pluralize(folderName)) ?? createElement("Folder", pluralize(folderName), element.id);
@@ -16,9 +24,7 @@ async function execute(element: IElementApi, domainClass?: IElementApi) {
 
     const resultDto = cqrsCrud.createCqrsResultTypeDto(entity, folder);
 
-    if (owningEntity == null || !privateSettersOnly) {
-        convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsCreateCommand(entity, folder, primaryKeys));
-    }
+    convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsCreateCommand(entity, folder, primaryKeys));
 
     if (hasPrimaryKey) {
         convertToAdvancedMapping.convertQuery(cqrsCrud.createCqrsFindByIdQuery(entity, folder, resultDto));
@@ -35,7 +41,7 @@ async function execute(element: IElementApi, domainClass?: IElementApi) {
         convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsCallOperationCommand(entity, operation, folder));
     }
 
-    if (hasPrimaryKey && (owningEntity == null || !privateSettersOnly)) {
+    if (hasPrimaryKey) {
         convertToAdvancedMapping.convertCommand(cqrsCrud.createCqrsDeleteCommand(entity, folder));
     }
 
@@ -43,6 +49,11 @@ async function execute(element: IElementApi, domainClass?: IElementApi) {
     diagramElement.loadDiagram();
     const diagram = getCurrentDiagram();
     diagram.layoutVisuals(folder, null, true);
+}
+
+
+function hasConstructor(entity: MacroApi.Context.IElementApi): boolean {
+    return entity.getChildren("Class Constructor").length > 0;
 }
 
 /**