+++ DesignerMacros/src/services-crud/create-crud-advanced-mapping/create-crud-macro-advanced-mapping.ts
@@ -8,6 +8,14 @@ async function execute(element: IElementApi) {
     });
     if (!entity) { return; }
 
+    if (privateSettersOnly && !hasConstructor(entity)) {
+        await dialogService.error(
+`Unable to create Service Operations.
+
+Private Setters are enabled with no constructor present on entity '${entity.getName()}'. In order to create Service Operations for that entity, either disable private setters or model a constructor element and try again.`);
+        return;
+    }
+
     const serviceName = `${toPascalCase(pluralize(DomainHelper.ownerIsAggregateRoot(entity) ? DomainHelper.getOwningAggregate(entity).getName() : entity.getName()))}Service`;
     const existingService = element.specialization == "Service" ? element : package.getChildren("Service").find(x => x.getName() == pluralize(serviceName));
     const service = existingService ?? createElement("Service", serviceName, package.id);
@@ -42,6 +50,10 @@ async function execute(element: IElementApi) {
     service.getChildren("Operation").forEach(operation => {
         convertToAdvancedMapping.convertOperation(operation, entity);
     })
+}
+
+function hasConstructor(entity: MacroApi.Context.IElementApi): boolean {
+    return entity.getChildren("Class Constructor").length > 0;
 }
 
 /**