+++ Modules/Intent.Modules.Modelers.Services.DomainInteractions/release-notes.md
@@ -2,6 +2,8 @@
 
 - Improvement: Changed `Default Mapping Mode` to be `Advanced` by default.
 - Improvement: Added setting `Default Query Implementation`, allowing for configuration of how `Query Entity Actions` are implemented.
+- Improvement: Create CRUD Scripts will inform the user when the Create operation couldn't be created due to `Ensure Private Property Setters` being enabled and no Constructor being present on the Entity.
+- Fixed: Create CRUD Scripts were breaking on `Ensure Private Property Setters`.
 
 ### Version 2.0.1